feat: 功能优化

+++ source-downloader-core/src/main/kotlin/xyz/shoaky/sourcedownloader/api/qbittorrent/TorrentInfo.kt
@@ -5,8 +5,8 @@ import com.fasterxml.jackson.annotation.JsonProperty
 //TODO 补充字段
 data class TorrentInfo(
     @JsonProperty("amount_left")
-    val amountLeft: Int,
+    val amountLeft: Long,
     val hash: String,
-    val size: Int,
+    val size: Long,
     val progress: Float
 )
